refactor : create camera module
