+++ app/src/main/java/com/kalex/dogescollection/di/appModule.kt
@@ -7,7 +7,7 @@ import com.kalex.dogescollection.core.api.DogsApi
 import com.kalex.dogescollection.core.common.Constants
 import com.kalex.dogescollection.dogList.model.repository.DogRepository
 import com.kalex.dogescollection.dogList.model.repository.DogRepositoryImpl
-import com.kalex.dogescollection.tensorflow.Classifier
+import tensorflow.Classifier
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -45,7 +45,10 @@ object ClassifierModule {
     }
     @Provides
     @ActivityScoped
-    fun provideClassifier(TfLiteModel: MappedByteBuffer, @ActivityContext context: Context): Classifier {
-        return Classifier(TfLiteModel, FileUtil.loadLabels(context , Constants.LABEL_PATH))
+    fun provideClassifier(TfLiteModel: MappedByteBuffer, @ActivityContext context: Context): tensorflow.Classifier {
+        return tensorflow.Classifier(
+            TfLiteModel,
+            FileUtil.loadLabels(context, Constants.LABEL_PATH)
+        )
     }
 }
