+++ camera/src/main/res/layout/fragment_camera.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".camera.CameraFragment">
+    tools:context="camera.CameraFragment">
 
 <com.google.android.material.card.MaterialCardView
     android:layout_width="match_parent"
