+++ settings.gradle
@@ -19,3 +19,4 @@ dependencyResolutionManagement {
 rootProject.name = "DogesCollection"
 include ':app'
 include ':core'
+include ':camera'
